requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.1.4

# Testing
mypy==0.761
pytest==5.3.4
attrs==19.1.0
pytest-sugar==0.9.2
pytest-django==3.8.0
factory-boy==2.11.1

# Code quality
flake8==3.6.0